+++ src/Eventbrite.php
@@ -71,7 +71,7 @@ class Eventbrite extends Plugin
    *
    * @var string
    */
-  public $schemaVersion = '1.0.0';
+  public $schemaVersion = '1.0.1';
   public $hasCpSettings = true;
   public $hasCpSection = true;